Improved performance by using components
Until now, the Score header used to be in the parent, and the Playing state used to be in the parent. So when the game stopped, the Score header rendered as well.

I fixed this by moving all the states (except the score-state) to a new Quiz component, and by doing so the parent component did not render the Score header when the Playing-state changed to false.
I also moved the Score header to its own component.